Transform async functions with regenerator
Remove `transform-async-to-generator`, which to my understanding is meant
to be used in environments that support generators natively.

Because we're compiling generators to ES5 anyway, we can simply use
`regenerator` to transform async functions too, which results in
slightly simpler output and only uses the regenerator runtime instead
regenerator runtime + _asyncToGenerator Babel helper.