Swap out babel-plugin-syntax-async-functions for fast-async
fast-async is based on nodent, a generator-optional implementation of async/await support.
By default, it does not use generators, which saves the usage of regenerator if generator functions are not otherwise used.
It is also significantly faster, with order-of-magnitude perf gains on mobile browsers in particular.
It also reduces file size by 1.44KB in a simple test case.