Add a more prominent link to the guide
As a contributor, I know where to find the documentation/guide since
know the codebase, but I've seen people miss the link to it in the
README, because it's not very prominent. Let's try to make it more
discoverable by including a link to it right on top of the page.