Fix script name to open chrome
A previous commit renamed the apple script to open chrome from
`chrome.applescript` to `openChrome.applescript`. That created
a minor bug. Even when chrome was open with the client app,
`npm start` would open the client app in a new Safari tab on
macOS instead of re-using the open tab in chrome.