Customize the port for openBrowser
Many times in local development, the port the app is listening on is not the same as the "public" port that the app is being locally tested on, particularly when the app consists of multiple, individually deployable modules intended to run under the same domain, some of which are data APIs.  In this case, it's inconvenient to have `start.js` open the browser to the port the app is listening on instead of the public port.

This file change supports the definition of a `PUBLIC_PORT` environment variable that will override the listening port when the `openBrowser` function is called.