Offline-First Considerations: index.html caching
If caches for service-worker.js and index.html get out of sync
then an old index.html might get populated in the cache
for a new service-worker.js. That does not get recovered
until a new service-worker.js is deployed. This is something
that is good to consider when deploying your create-react-app site.