Use a flag to watch stdin
The kitchensink e2e tests were failing because in those tests the
`npm start` command is sent to the background. The shell will then halt
any background job that tries to read from the standard input, and since
we are watching standard input to determine if we should exit or not the
dev server would immediately exit.

The workaround is to provide run the dev-server as
`npm start </dev/zero &` instead, which provides a different standard
input that will not close immediately. This, alas, doesn't work on
Windows and might change the expected behaviour for some people that are
used to start the dev-server as a background job.

For this reason a --watch-stdin flag has been added to only enable this
behaviour on request for the few applicable use cases.

A note has also been added in the Troubleshooting section to help any
final user that might start the dev-server as a background job.