Make the threshold whether assets should be inlined confiurable for prod builds
For sites which have a large number of images, each of which has a lower probability of being seen,
combined with assets which rarely change, the other situation was inconvenient. Users were forced
to download the same assets over and over again for unrelated changes.

Apart from that, since each image has a low probability of being loaded, aggressively loading all
images is counterproductive.

By making the threshold configurable, but respecting the same default value, developers can use
techniques as lazyloading images when applicable (which cannot be done if they are part of the bundle)