Create bundle stats with new build script
As discussed in #1858, we'd like to make it easy for users to view
bundle stats without ejecting.

This PR adds a new script that, along with building the site for
production, creates a `build-stats.json` file. This file can be
used with tools like Webpack Bundle Analyzer or Webpack Visualizer,
to help the user optimize their bundles.

Conflicts:
	package.json
	packages/react-scripts/package.json