Add directory details to packages/* package.json
Specifying the directory as part of the repository field in a
package.json allows third party tools to provide better support when
working with monorepos. For example, it allows them to correctly
construct a commit diff for a specific package.

This format was accepted by npm in [npm/rfcs#19](https://github.com/npm/rfcs/pull/19).