Fix hanging start
CRA 3 opts-in to a new webpack 5 feature early to save memory: `futureEmitAssets`
https://github.com/facebook/create-react-app/pull/6696
This optimization only stores the file sizes during webpack compilation, no longer storing the original source
https://github.com/webpack/webpack/pull/8642
This is incompatible with our webpack-dev-server customization setting of `writeToDisk`
https://github.com/webpack/webpack/pull/8642#issuecomment-455342804
which tries to read the source after emit, which throws an exception, but that error is swallowed and the `npm start` just hangs.
Before we update to webpack 5, where this feature will be the default, we need to stop writing to disk, or find another way to do it than the `writeToDisk` flag.