Exit from startBrowserProcess when no browser is provided.
In cases where no browser is set, the open library fails with an ENOENT.
This means that on Windows 10 with Windows Subsystem For Linux, if the user
has not set the Windows paths in their PATH, this will fail.
Thus, if no browser is provided, we return early to avoid an (inevitable) crash.

This solves #7251.