Disable babel-loader's cacheCompression
`cacheCompression` is an enabled-by-default flag in babel-loader,
which gzips the babel-loader cache.

Most projects do not actually benefit from cache compression,
especially in production where builds often happen in a CI environment
where memory is precious and disk space is not.

Furthermore, having it disabled in dev and enabled in prod means caching
won't be shared between dev and prod.

Finally, the disk space savings are not that great. Most React projects
will have a lot of small files (one file per component). babel-loader
caches each file as a unit, so the compression overhead increases when
the project is mainly lots of small files.

For these reasons, we're disabling cache compression.

Closes #6846