fix: add template support for homepage, browserslist properties
This change updates the init script to allow for the homepage property to be set by a template. The homepage field with a value of "." is used when running applications at non-root of a domain. Additionally this adds support for templates to customize the browserlist property. This is useful specifically when a template includes IE support out of the box.