Add optional support for custom postcss plugins
This is accomplished via postcss.config.js, which will be searched for if `plugins` is not specified as an option to `postcss-loader`

Postcss plugins are required to use Tailwind CSS, a popular styling framework which leverages a custom module import syntax.

This overlaps slightly with #8331, but takes a non-breaking change approach.