Fixes author credit in changelog
I created the PR referenced but it was @jakehasler 's original work. See the description of PR #7320:

This is the same functionality as PR #3430 originally submitted by @jakehasler
The original PR broke the CI. It went unaddressed, went stale, and was automatically closed.
Like Jake, my team is integrating CRA with an elixir-phoenix app and the CRA isn't getting closed when the main process is killed.