fix: improve DefinePlugin config
This configuration is used in webpack's DefinePlugin which prefers keys
that look like `process.env.REACT_APP_MY_ENV_VAR` instead of a nested
object: https://webpack.js.org/plugins/define-plugin/

Prior to this change, webpack would replace all occurrences of
`process.env` with a JS object containing every key in the environment
that begins with `REACT_APP`.

The expected behavior is that a reference to
`process.env.REACT_APP_MY_ENV_VAR` is replaced with the value of just
that environment variable.