Fix broken links
The shortened google links are no longer working. 

They now take you to an empty page that redirects you to another page with many links, so you have to read and click to yet another page to get to the docs you need. This is a pain and doesn't seem in the spirit of CRA.

The actual links aren't actually much shorter than the shortened ones, so let's just link to the actual documentation and safe everyone some time.

Related: https://github.com/facebook/create-react-app/issues/8188