feat: support custom hosts ending with .localhost
This makes `react-scripts start` just work if the user specified a custom
HOST environment variable that ends with .localhost.
The host is internally converted to "localhost" to avoid a failed DNS
resolution while starting the dev server but is still used in full to
print the URL to the terminal and open the browser.