docs: Fix confusing section reference
The "next section" mentioned in the note on an alternative way of handling static assets is no longer the section on using the public folder (a section on loading .graphql files is now the next section). This clarifies the previous wording and includes an explicit link to the relevant section.