feat: remove cached lockfile
To go along with #11474, we no longer need the cached lockfile after unpinning all the dependencies.

This PR removes all references to it in the creation of new projects and the tests. It also refactors the integration test to assert that all files exist and no extra ones are being written.