Add README install for Yarn 2+ using 'yarn dlx'
This commit brings the README up to date with yarn latest.

See <https://yarnpkg.com/>

Yarn 3 is the latest stable version of `yarn`,
such as installed by running this command:

```sh
$ yarn set version latest
```

Yarn 2 introduced the new command 'yarn dlx' which downloads and executes
a package in a temporary environment. This is the preferred way to work with
Yarn 2 and Yarn 3, as explained on the Yarn website page about 'yarn dlx'.

See <https://yarnpkg.com/cli/dlx>

Yarn 1 is helpful to still have in the instructions because some developers
still use it for historical reasons, such as needing to manage a third-party 
package that has not been updated to work with Yarn 2 or Yarn 3.
However, the Yarn 1 instructions do not work correctly with Yarn latest.