Update App.js
Due to the general importance of digital accessibility in general and even more in the future I would be very glad if the REACT Team could improve this template to create valid markup. One of the major issues I see, when I support development teams working with react very often is "missing <main>-landmark".

Besides that to raise the awareness for accessibility it would be awesome to add some links to basic accessibility documentation. THANKS ind advance.