Adjust default build directory for Netlify deploy
Netlify automatically changes the "Publish directory" to `build` when it detects
a `create-react-app` application. But we are using a scoped module
@upleveled/react-scripts, which it cannot figure out. So Netlify falls back to
its default of `dist/`, which causes all apps created using
`pnpm create @upleveled/react-app ...` to fail to deploy (404 page - since the
dist/ folder does not exist).

In this commit we're setting the default build path to `dist`, to be compatible
with the default Netlify "Publish directory"