Deprecate Create React App officially by changing the README, and adding a message on init (#17003)
It's probably time to make this project document its status as being
deprecated and not recommended for production usage.

To change it:

- I opted to add a header to the README saying its over and you should
go look at https://react.dev/learn/start-a-new-react-project

- I left a note saying that if you are following, it is maybe worth
carrying on. While I hear react 19 doesn't work with CRA, I wouldn't be
surprised that a good chunk of tutorials would still work. Open to being
a bit more hard-lined there but there was a lot of great resources for
learning react in that era and it seems like a waste to be making people
stop early?

- I added a message inside the CLI, it shows once and says "don't use
this, use the stuff in
https://react.dev/learn/start-a-new-react-project"

---------

Co-authored-by: Ricky <rickhanlonii@gmail.com>
Co-authored-by: Rick Hanlon <rickhanlonii@fb.com>